graph: move node implementations to separate file
