+++ _msvc/zimg/zimg.vcxproj
@@ -331,6 +331,7 @@
     <ClInclude Include="..\..\src\zimg\graph\copy_filter.h" />
     <ClInclude Include="..\..\src\zimg\graph\filtergraph.h" />
     <ClInclude Include="..\..\src\zimg\graph\filtergraph2.h" />
+    <ClInclude Include="..\..\src\zimg\graph\graphnode.h" />
     <ClInclude Include="..\..\src\zimg\graph\graphbuilder.h" />
     <ClInclude Include="..\..\src\zimg\graph\graphbuilder2.h" />
     <ClInclude Include="..\..\src\zimg\graph\image_filter.h" />
@@ -458,6 +459,7 @@
     <ClCompile Include="..\..\src\zimg\graph\filtergraph2.cpp" />
     <ClCompile Include="..\..\src\zimg\graph\graphbuilder.cpp" />
     <ClCompile Include="..\..\src\zimg\graph\graphbuilder2.cpp" />
+    <ClCompile Include="..\..\src\zimg\graph\graphnode.cpp" />
     <ClCompile Include="..\..\src\zimg\resize\filter.cpp" />
     <ClCompile Include="..\..\src\zimg\resize\resize.cpp" />
     <ClCompile Include="..\..\src\zimg\resize\resize_impl.cpp" />
